# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Agustinos.tv API",
        "version": "3.20.01"
    },
    "servers": [
        {
            "url": "http://api.agustinos.tv"
        }
    ],
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Registra un nuevo usuario",
                "description": "Registra un nuevo usuario",
                "operationId": "App\\Http\\Controllers\\AuthController::register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Email del usuario",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Contraseña del usuario",
                                        "type": "string"
                                    },
                                    "username": {
                                        "description": "Nombre de Usuario",
                                        "type": "string"
                                    },
                                    "acepto_avisos_contenido": {
                                        "description": "Acepta el envío de notificaciones y emails sobre cambios en el contenido de Agustinos.TV (notificaciones push por ej)",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Contraseña modificada con éxito"
                    },
                    "404": {
                        "description": "El token no es válido o ha caducado"
                    },
                    "401": {
                        "description": "Usuario no encontrado"
                    }
                }
            }
        },
        "/api/auth/ask-activate/{email}": {
            "get": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Reenviar Activación",
                "description": "Reenviar Activación",
                "operationId": "App\\Http\\Controllers\\AuthController::PedirActivacion",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Email del usuario",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Notificación enviada"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    },
                    "409": {
                        "description": "Cuenta ya activada"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Login",
                "description": "Login",
                "operationId": "App\\Http\\Controllers\\AuthController::login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Email de Usuario a autenticar",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Contraseña del Usuario a autenticar",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario autenticado correctamente"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    },
                    "401": {
                        "description": "Usuario no activado | Datos incorrectos (Unauthorized)"
                    }
                }
            }
        },
...
        "/api/auth/user": {
            "get": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Obtener el usuario conectado",
                "description": "Obtener el usuario conectado",
                "operationId": "App\\Http\\Controllers\\AuthController::user",
                "responses": {
                    "200": {
                        "description": "Objeto del usuario conectado"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                }
            }
        },
        "/api/auth/password-recovery/ask": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Pedir recuperación de contraseña",
                "description": "Pedir recuperación de contraseña",
                "operationId": "App\\Http\\Controllers\\AuthController::PeticionRecuperarContrasenya",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Email del usuario",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Petición enviada"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                }
            }
        },
        "/api/auth/password-recovery/update": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Realizar cambio de contraseña",
                "description": "Realizar cambio de contraseña",
                "operationId": "App\\Http\\Controllers\\AuthController::ActualizarContrasenya",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "token": {
                                        "description": "Token de la Petición de cambio de contraseña",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Nueva contraseña",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Contraseña modificada con éxito"
                    },
                    "404": {
                        "description": "El token no es válido o ha caducado"
                    },
                    "401": {
                        "description": "Usuario no encontrado"
                    },
                    "409": {
                        "description": "Faltan datos"
                    }
                }
            }
        },
        "/api/auth/activate/{token}": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Activar una cuenta",
                "description": "Activar una cuenta",
                "operationId": "App\\Http\\Controllers\\AuthController::Activar",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "token": {
                                        "description": "Email de Usuario a autenticar",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                }
            }
        },
...
    }
}